Introducing SimpleDMAController
We are testing the DMA operation when a packet with data is read through
the ICENet data path and the data is written to memory and read back
from memory using DMA and the packet is setn back. intial effort was
done to use streaming interfaces introduced by IceNet. But this version
use custom DMA modules to achieve the target.