Improved Recusrive Doubling Module with DMA
Recursive Doubling module is imporved with DMA where the incoming
packets are stored into memory using DMA and the processed packets are
stored in the memory as well. While the packet sizes handled here are
the same size, this is done to enable larger data sizes to be processed.
This is an intial commit of a working design, which will be further
improved to optimize the state machine.